- minor refactor of the initializeRunStore logic - extract common logic
